requirements.txt
python-telegram-bot==13.7.0
flask==2.0.1
Werkzeug==2.0.3
gunicorn==20.1.0